requirements.txt
boto==2.38.0
Django==1.8.4
gunicorn==19.3.0
MySQL-python==1.2.5
python-dateutil==2.4.2
six==1.10.0
wheel==0.24.0
whitenoise==2.0.6
